+++ src/backend/core/api/viewsets.py
@@ -46,11 +46,14 @@ from core.recording.event.exceptions import (
     InvalidFileTypeError,
     ParsingEventDataError,
 )
-from core.recording.event.notification import notification_service
 from core.recording.event.parsers import get_parser
 from core.recording.services.metadata_collector import (
     MetadataCollectorException,
     MetadataCollectorService,
+)
+from core.recording.services.recording_events import (
+    RecordingEventsService,
+    RecordingNotSavableError,
 )
 from core.recording.worker.exceptions import (
     RecordingStartError,
@@ -972,24 +975,15 @@ class RecordingViewSet(
         except models.Recording.DoesNotExist as e:
             raise drf_exceptions.NotFound("No recording found for this event.") from e
 
-        if not recording.is_savable():
+        # Save recording
+        recording_events_service = RecordingEventsService()
+        try:
+            recording_events_service.handle_complete(recording)
+        except RecordingNotSavableError:
             raise drf_exceptions.PermissionDenied(
                 f"Recording with ID {recording_id} cannot be saved because it is either,"
                 " in an error state or has already been saved."
-            )
-
-        # Attempt to notify external services about the recording
-        # This is a non-blocking operation - failures are logged but don't interrupt the flow
-        notification_succeeded = notification_service.notify_external_services(
-            recording
-        )
-
-        recording.status = (
-            models.RecordingStatusChoices.NOTIFICATION_SUCCEEDED
-            if notification_succeeded
-            else models.RecordingStatusChoices.SAVED
-        )
-        recording.save()
+            ) from None
 
         return drf_response.Response(
             {"message": "Event processed."},